backend/requirements.txt
Django==4.2.23
psycopg2-binary
python-decouple
django-crispy-forms
crispy-bootstrap4
crispy-bootstrap5
django-filter
djangorestframework
markdown
django-environ
drf-yasg
gunicorn
whitenoise
django-cors-headers
factory-boy
pytest-django
django-debug-toolbar
Pillow
